# backend/internal/api/docs/configer_swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Write-back-native GitOps configuration management. Configer discovers an existing repository's layout, exposes a parameter x instance grid over the repository's OWN files, and surgically edits them; every change flows to Git as commits, branches, pull requests, and merges. `.configer/` holds metadata only.\n\n## Repository scoping\nEvery resource endpoint exists in two shapes: the unscoped `/api/...` form acts on the default (first-connected) repository for single-repo deployments and older clients, and the identical routes are available per repository under `/api/repos/{repoId}/...`. Only the unscoped forms are documented here; to target a specific repository prefix any path with `/api/repos/{repoId}`.\n\n## Errors\nEvery 4xx/5xx response uses one envelope: a stable machine-readable `code`, a human-readable `error`, a `requestId` (also returned as the `X-Request-ID` header) that correlates to server logs, and - for validation failures - a `fields` array. Branch on `code`, never on the message text.\n\n## Correlation\nSend `X-Request-ID` to have it echoed and used as the correlation id; otherwise the server generates one. It is returned on every response.",
        "title": "Configer API",
...
        "contact": {
            "name": "Configer",
...
        },
...
    },
    "basePath": "/",
    "paths": {
...
        "/api/parameters": {
            "post": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Create a new catalog parameter. Bindings may be empty for a design-phase parameter (attached later), but a half-specified binding is rejected. Committed directly with attribution.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grid \u0026 parameters"
                ],
                "summary": "Create a parameter",
                "parameters": [
                    {
                        "description": "{param: Parameter, author?: string}",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Parameter"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "URL of the created parameter"
                            }
                        }
                    },
                    "400": {
                        "description": "Malformed body, missing name, or half-specified binding",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "409": {
                        "description": "A parameter with that id already exists",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/parameters/{id}": {
            "get": {
                "description": "Full metadata for one catalog parameter: type, scope, category, bindings, validation and version state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grid \u0026 parameters"
                ],
                "summary": "Get a parameter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parameter id (slug)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Parameter"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Patch a parameter's type, validation, display name, description, category, scope, secret flag, default, or file bindings. Nil fields are left unchanged. Committed directly to the working branch with attribution.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grid \u0026 parameters"
                ],
                "summary": "Update a parameter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parameter id (slug)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Catalog revision the edit is based on (from a read's ETag)",
                        "name": "If-Match",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Partial parameter patch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Parameter"
                        }
                    },
                    "400": {
                        "description": "Malformed body or half-specified binding",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "404": {
                        "description": "Unknown parameter",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "412": {
                        "description": "Stale revision; reload and reapply",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "422": {
                        "description": "Unknown validation preset",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "428": {
                        "description": "Missing If-Match",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieSession": []
                    }
                ],
                "description": "Retire a parameter everywhere: the catalog entry is removed and the bound key/element is deleted from every real file it lives in. Committed directly with attribution.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grid \u0026 parameters"
                ],
                "summary": "Delete a parameter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Parameter id (slug)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OKResponse"
                        }
                    },
                    "404": {
                        "description": "Unknown parameter",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "409": {
                        "description": "Deletion failed",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the stable, machine-readable classifier a client branches on.",
                    "type": "string",
                    "example": "not_found"
                },
                "error": {
                    "description": "Error is a human-readable, plain-language message safe to show a user.\nIt never contains git jargon, stack traces, credentials, or internal\nhostnames.",
                    "type": "string",
                    "example": "parameter not found"
                },
                "fields": {
                    "description": "Fields carries per-field validation failures (422 only).",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.FieldError"
                    }
                },
                "requestId": {
                    "description": "RequestID correlates this response with the access log and the\nX-Request-ID header, so an operator can find the exact request.",
                    "type": "string",
                    "example": "3f9a1c2b7e5d0a84"
                }
            }
        },
...
        "api.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "port"
                },
                "message": {
                    "type": "string",
                    "example": "must be between 1 and 65535"
                }
            }
        },
...
        "api.OKResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "model.Binding": {
            "type": "object",
            "properties": {
                "file": {
                    "type": "string"
                },
                "format": {
                    "description": "yaml | json | xml",
                    "type": "string"
                },
                "layer": {
                    "description": "Layer overrides the inferred precedence layer (\"base\" or \"instance\").",
                    "type": "string"
                },
                "line": {
                    "description": "Line is the 1-based source line the value lives on, for display in the\nonboarding proposal. It is NEVER persisted (line numbers drift), only\ncarried through the discovery JSON - hence yaml:\"-\".",
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "model.ParamType": {
            "type": "string",
            "enum": [
                "string",
                "integer",
                "boolean",
                "number",
                "enum",
                "ipv4",
                "ipv6",
                "cidr",
                "hostname",
                "port",
                "email",
                "url",
                "mac",
                "cpu",
                "memory",
                "duration",
                "percentage",
                "list"
            ],
            "x-enum-varnames": [
                "TypeString",
                "TypeInteger",
                "TypeBoolean",
                "TypeNumber",
                "TypeEnum",
                "TypeIPv4",
                "TypeIPv6",
                "TypeCIDR",
                "TypeHostname",
                "TypePort",
                "TypeEmail",
                "TypeURL",
                "TypeMAC",
                "TypeCPU",
                "TypeMemory",
                "TypeDuration",
                "TypePercentage",
                "TypeList"
            ]
        },
        "model.Parameter": {
            "type": "object",
            "properties": {
                "bindings": {
                    "description": "Bindings are the real-file locations this parameter's value lives at.\nA deduplicated parameter (the same logical setting appearing in several\nfiles) carries one binding per location; an edit fans out to all of\nthem. A design-phase parameter has no bindings yet.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Binding"
                    }
                },
                "category": {
                    "type": "string"
                },
                "default": {
                    "description": "Default is metadata: the value shown (and written on first set) when no\nbound file carries the key. It is never rendered anywhere by itself."
                },
                "dependsOn": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "derived": {
                    "description": "Derived, when set, is a computed default expressed in terms of another\nparameter: \"{other-id}\" copies that parameter's effective value for the\nsame instance, optionally with an integer offset (\"{other-id}+1\"). It is\nresolved read-only (never written) and any real file value overrides it,\nso it stays true to the write-back model: a suggestion, not a stored value.",
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "itemType": {
                    "description": "ItemType is the element type when Type is \"list\".",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.ParamType"
                        }
                    ]
                },
                "name": {
                    "type": "string"
                },
                "observed": {
                    "description": "Observed carries the value discovery read from each instance's files\n(instance name -\u003e value), so the onboarding proposal can preview the grid.\nIt is display-only: never persisted to .configer (yaml:\"-\"), only carried\nthrough the discovery JSON.",
                    "type": "object",
                    "additionalProperties": {}
                },
                "scope": {
                    "$ref": "#/definitions/model.Scope"
                },
                "secret": {
                    "type": "boolean"
                },
                "source": {
                    "description": "Source, when set, maps this parameter to a key in an external source\n(see .configer/sources.yaml). The source's value is never applied\nsilently: it surfaces as an \"incoming change\" the reviewer accepts into a\ndraft, so the write-back model stays intact (a repo file value is only\nchanged by a reviewed commit). Analogous to Derived, but the value comes\nfrom another system (another Git repo, a secret store) instead of another\nparameter.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.SourceRef"
                        }
                    ]
                },
                "type": {
                    "$ref": "#/definitions/model.ParamType"
                },
                "validation": {
                    "$ref": "#/definitions/model.Validation"
                },
                "versionDeprecated": {
                    "type": "string"
                },
                "versionIntroduced": {
                    "description": "VersionIntroduced/Deprecated drive version-aware cell state in the grid.",
                    "type": "string"
                }
            }
        },
        "model.Scope": {
            "type": "string",
            "enum": [
                "instance",
                "global"
            ],
            "x-enum-varnames": [
                "ScopeInstance",
                "ScopeGlobal"
            ]
        },
        "model.SourceRef": {
            "type": "object",
            "properties": {
                "instance": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "sourceId": {
                    "type": "string"
                }
            }
        },
        "model.Validation": {
            "type": "object",
            "properties": {
                "atLeast": {
                    "description": "AtLeast / AtMost name another parameter (by id) whose effective value at\nthe same instance bounds this one: a resource limit must be AtLeast its\nrequest, a request AtMost its limit. The comparison is quantity-aware\n(CPU millicores, memory bytes, otherwise numeric), so \"1\" \u003e= \"500m\" and\n\"1Gi\" \u003e= \"512Mi\" compare correctly. Enforced on write when the related\nparameter can be resolved.",
                    "type": "string"
                },
                "atMost": {
                    "type": "string"
                },
                "enum": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "max": {
                    "type": "number"
                },
                "maxItems": {
                    "type": "integer"
                },
                "maxLength": {
                    "type": "integer"
                },
                "min": {
                    "type": "number"
                },
                "minItems": {
                    "type": "integer"
                },
                "minLength": {
                    "type": "integer"
                },
                "pattern": {
                    "type": "string"
                },
                "preset": {
                    "description": "id of a predefined rule",
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "schemaRef": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieSession": {
            "description": "Session cookie set by the GitHub OAuth login flow (`GET /api/auth/login` -\u003e callback). Only required when the deployment has OAuth configured; single-user deployments need no authentication.",
            "type": "apiKey",
            "name": "configer_session",
            "in": "cookie"
        }
    }
}